+++ dev/Network.js
@@ -172,6 +172,8 @@ class Network {
 
     joinLayer (layer, layerIndex) {
 
+        layer.net = this
+
         layer.activation = this.activation
         layer.adaptiveLR = this.adaptiveLR
         layer.activationConfig = this.activationConfig
@@ -235,7 +237,14 @@ class Network {
         }
     }
 
-    train (dataSet, {epochs=1, callback, log=true}={}) {
+    train (dataSet, {epochs=1, callback, log=true, miniBatchSize=1}={}) {
+
+        this.miniBatchSize = typeof miniBatchSize=="boolean" && miniBatchSize ? dataSet[0].expected.length : miniBatchSize
+
+        if (log) {
+            console.log(`Training started. Epochs: ${epochs} Batch Size: ${this.miniBatchSize}`)
+        }
+
         return new Promise((resolve, reject) => {
             
             if (dataSet === undefined || dataSet === null) {
@@ -266,17 +275,21 @@ class Network {
             const doIteration = () => {
                 
                 if (!dataSet[iterationIndex].hasOwnProperty("input") || (!dataSet[iterationIndex].hasOwnProperty("expected") && !dataSet[iterationIndex].hasOwnProperty("output"))) {
-                    return reject("Data set must be a list of objects with keys: 'input' and 'expected' (or 'output')")
+                    return void reject("Data set must be a list of objects with keys: 'input' and 'expected' (or 'output')")
                 }
-
-                this.resetDeltaWeights()
 
                 const input = dataSet[iterationIndex].input
                 const output = this.forward(input)
                 const target = dataSet[iterationIndex].expected || dataSet[iterationIndex].output
 
                 this.backward(target)
-                this.applyDeltaWeights()
+
+                if (++iterationIndex%this.miniBatchSize==0) {
+                    this.applyDeltaWeights()
+                    this.resetDeltaWeights()
+                } else if (iterationIndex >= dataSet.length) {
+                    this.applyDeltaWeights()
+                }
 
                 const iterationError = this.cost(target, output)
                 const elapsed = Date.now() - startTime
@@ -291,7 +304,6 @@ class Network {
                 }
 
                 this.iterations++
-                iterationIndex++
 
                 if (iterationIndex < dataSet.length) {
                     setTimeout(doIteration.bind(this), 0)
@@ -317,6 +329,7 @@ class Network {
                 }
             }
 
+            this.resetDeltaWeights()
             doEpoch()
         })
     }
@@ -347,7 +360,7 @@ class Network {
 
                 if (iterationIndex < testSet.length) {
                     setTimeout(testInput.bind(this), 0)
-                    
+
                 } else {
                     const elapsed = Date.now() - startTime